AIComplianceMonitoring/agents/ui_agent/requirements_ui.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
SQLAlchemy==2.0.31
bcrypt==4.1.3
Werkzeug==3.0.3
redis==5.0.7
Flask-Limiter==3.6.0
PyJWT==2.8.0
email-validator==2.2.0
